+++ CoreComponents/Core/Misc/InvalidTypeException.cs
@@ -1,0 +1,142 @@
+﻿/*
+InvalidTypeException.cs
+
+This file is part of Morgan's CLR Advanced Runtime (MCART)
+
+Author(s):
+     César Andrés Morgan <xds_xps_ivx@hotmail.com>
+
+Copyright (c) 2011 - 2018 César Andrés Morgan
+
+Morgan's CLR Advanced Runtime (MCART) is free software: you can redistribute it
+and/or modify it under the terms of the GNU General Public License as published
+by the Free Software Foundation, either version 3 of the License, or (at your
+option) any later version.
+
+Morgan's CLR Advanced Runtime (MCART) is distributed in the hope that it will be
+useful, but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General
+Public License for more details.
+
+You should have received a copy of the GNU General Public License along with
+this program. If not, see <http://www.gnu.org/licenses/>.
+*/
+
+#region Configuración de ReSharper
+
+// ReSharper disable MemberCanBeProtected.Global
+// ReSharper disable MemberCanBePrivate.Global
+// ReSharper disable UnusedMember.Global
+// ReSharper disable UnusedAutoPropertyAccessor.Global
+
+#endregion
+
+using System;
+using System.Runtime.Serialization;
+using TheXDS.MCART.Annotations;
+using TheXDS.MCART.Resources;
+
+namespace TheXDS.MCART.Exceptions
+{
+    /// <inheritdoc />
+    /// <summary>
+    ///     Excepción que se produce al hacer referencia a un tipo inválido
+    /// </summary>
+    [Serializable]
+    public class InvalidTypeException : OffendingException<Type>
+    {
+        private static string Msg()=> Strings.XIsInvalid(Strings.TheType);
+        private static string Msg([NotNull]Type type) => Strings.XIsInvalid(Strings.XYQuotes(Strings.TheType, type.FullName) );
+
+        /// <inheritdoc />
+        /// <summary>
+        ///     Inicializa una nueva instancia de la clase <see cref="T:TheXDS.MCART.Exceptions.InvalidTypeException" />.
+        /// </summary>
+        public InvalidTypeException() : base(Msg())
+        {
+        }
+
+        /// <inheritdoc />
+        /// <summary>
+        ///     Inicializa una nueva instancia de la clase <see cref="T:TheXDS.MCART.Exceptions.InvalidTypeException" />.
+        /// </summary>
+        /// <param name="offendingType">Tipo que ha causado la excepción.</param>
+        public InvalidTypeException(Type offendingType) : base(Msg(offendingType), offendingType)
+        {
+        }
+
+        /// <inheritdoc />
+        /// <summary>
+        ///     Inicializa una nueva instancia de la clase <see cref="T:TheXDS.MCART.Exceptions.InvalidTypeException" />.
+        /// </summary>
+        /// <param name="message">Un <see cref="T:System.String" /> que describe a la excepción.</param>
+        public InvalidTypeException(string message) : base(message)
+        {
+        }
+
+        /// <inheritdoc />
+        /// <summary>
+        ///     Inicializa una nueva instancia de la clase <see cref="T:TheXDS.MCART.Exceptions.InvalidTypeException" />.
+        /// </summary>
+        /// <param name="message">Un <see cref="T:System.String" /> que describe a la excepción.</param>
+        /// <param name="offendingType">Tipo que ha causado la excepción.</param>
+        public InvalidTypeException(string message, Type offendingType) : base(message, offendingType)
+        {
+        }
+
+        /// <inheritdoc />
+        /// <summary>
+        ///     Inicializa una nueva instancia de la clase <see cref="T:TheXDS.MCART.Exceptions.InvalidTypeException" />.
+        /// </summary>
+        /// <param name="inner"><see cref="T:System.Exception" /> que es la causa de esta excepción.</param>
+        public InvalidTypeException(Exception inner) : base(Msg(), inner)
+        {
+        }
+
+        /// <inheritdoc />
+        /// <summary>
+        ///     Inicializa una nueva instancia de la clase <see cref="T:TheXDS.MCART.Exceptions.InvalidTypeException" />.
+        /// </summary>
+        /// <param name="message">Un <see cref="T:System.String" /> que describe a la excepción.</param>
+        /// <param name="inner"><see cref="T:System.Exception" /> que es la causa de esta excepción.</param>
+        public InvalidTypeException(string message, Exception inner) : base(message, inner)
+        {
+        }
+
+        /// <inheritdoc />
+        /// <summary>
+        ///     Inicializa una nueva instancia de la clase <see cref="T:TheXDS.MCART.Exceptions.InvalidTypeException" />.
+        /// </summary>
+        /// <param name="inner"><see cref="T:System.Exception" /> que es la causa de esta excepción.</param>
+        /// <param name="offendingType">Tipo que ha causado la excepción.</param>
+        public InvalidTypeException(Exception inner, Type offendingType) : base(Msg(offendingType), inner)
+        {
+        }
+
+        /// <inheritdoc />
+        /// <summary>
+        ///     Inicializa una nueva instancia de la clase <see cref="T:TheXDS.MCART.Exceptions.InvalidTypeException" />.
+        /// </summary>
+        /// <param name="message">Un <see cref="T:System.String" /> que describe a la excepción.</param>
+        /// <param name="inner"><see cref="T:System.Exception" /> que es la causa de esta excepción.</param>
+        /// <param name="offendingType">Tipo que ha causado la excepción.</param>
+        public InvalidTypeException(string message, Exception inner, Type offendingType) : base(message, inner, offendingType)
+        {
+        }
+
+        /// <inheritdoc />
+        /// <summary>
+        ///     Inicializa una nueva instancia de la clase
+        ///     <see cref="T:TheXDS.MCART.Exceptions.InvalidTypeException" />.
+        /// </summary>
+        /// <param name="info">
+        ///     El objeto que contiene la información de serialización.
+        /// </param>
+        /// <param name="context">
+        ///     La información contextual acerca del orígen o el destino.
+        /// </param>
+        protected InvalidTypeException(SerializationInfo info, StreamingContext context) : base(info, context)
+        {
+        }
+    }
+}
